addressing most of the comments
